Fix date in female condition file name
Was 2021-02-22, should be 2022-02-22. Note that including dates in data file names is problematic because every update breaks the data-munging script. It's also unnecessary when using version tracking such as Git. We should either drop the dates or modify the data script to ignore them by partial matching (but could be dangerous if multiple versions ever exist in target directory).