Fix error in Duncan Channel straying estimates and plots
I had used spawner_data_agg as the source for S_obs when doing these calculations and plots. The problem is that spawner_data_agg$pop is defined based on disposition not location, so almost no spawners are recorded for Duncan Creek (which is eventually dropped when constructing fish_data). This meant that the expanded estimates of returns to Duncan Creek by origin (in particular Duncan Channel) were actually less than the corresponding number of recoveries. The issue is resolved by starting with spawner_data instead of spawner_data_agg when constructing the data frames used in these quick and dirty plots.