Initial attempt at future hatchery scenarios
(1) No broodstock removal or hatchery smolt releases
(2) Maximum observed broodstock removal and hatchery smolt releases in each pop

Unexpected results: maximum spawners are reduced, not only in pops that are mined for broodstock but also (to a lesser degree) in those that have no broodstock take. This is not the result of Ricker overcompensation; it occurs in BH too. If releases are maintained while broodstock take is zero, results are largely as expected (higher floor for spawners, no effect on ceiling).

Seems like this must be a bug, but I have no idea what.