Refine updated vignette
Some previously undetected numerical mischief is causing S_origin, and thus p_HOS, to be nonzero (trace values on the order of 1e-7 to 1e-3) even when all hatchery M_obs are set to 0. Even stranger, this was not always the case (see history of p_HOS_fore_Ricker.png), yet nothing has changed in salmonIPM since then.

Need to figure this out, but in the meantime set future mean p_HOS under H0 to 0.