Reorganize vignette: Retrospective Models
The flow of the retrospective model fitting section has always been anchored by the multiplot, which shows hyperparameters and (some) states for the entire life cycle. This gets awkward when later subsections address observation models (fits of params / states to data) of the same life stages and transitions shown in the multiplot.

It makes more sense to organize the results around the life history and relevant applied questions. That means it's time to retire the multiplot, a cool figure that is still useful in interactive mode as a quick overview.

Retrospective Model subsections are reorganized, figures are updated and new ones added. The text still needs to be edited to flow with the new structure and to emphasize relevant themes.