Deep dive into mysterious nonzero future p_HOS under H0 scenario
The reason p_origin / p_HOS are nonzero even when all hatchery M_obs == 0 is that stan_data() converts 0 abundances to 1. (It does not flag them as missing, though it also changes NA to 1 and then ignores them.) This produces tiny trace amounts of S_H from the M_obs == 1. Hatchery M_obs is considered known w/o error in the current version of the model, but that will change once we build the broodstock-to-smolt stage (true hatchery smolts will be an unknown state M), so it's not worth worrying about now.